+++ sound/soc/sof/Makefile
@@ -26,7 +26,7 @@ snd-sof-of-objs := sof-of-dev.o
 
 snd-sof-ipc-flood-test-objs := sof-client-ipc-flood-test.o
 snd-sof-ipc-msg-injector-objs := sof-client-ipc-msg-injector.o
-snd-sof-probes-objs := sof-client-probes.o
+snd-sof-probes-objs := sof-client-probes.o sof-client-probes-ipc3.o
 
 snd-sof-nocodec-objs := nocodec.o
